# api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Lockdown app
  description: This is the api for the lockdown app
  termsOfService: http://swagger.io/terms/
  contact:
    name: Lock Down team
    email: nonem@swagger.io
    url: https://github.com/Lockdown-WirVsVirus/lockdown-app-backend
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://lockdown.api/api
paths:
  /leaverequest:
    post:
      description: Create a new Leave Request
      operationId: addLeaveRequest
      requestBody:
        description: Create a new LeaveRequest
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLeaveRequest'
      responses:
        '200':
          description: Leave Request response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /leaverequest/{id}:
    get:
      description: Returns a user based on a single ID
      operationId: getLeaveRequest
      parameters:
        - name: id
          in: path
          description: ID of the leave request
          required: true
          schema:
            type: integer
            format: int64
        - name: pin
          in: query
          description: Pin for the ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Leave request response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: deletes a LeaveRequest
      operationId: deletePet
      parameters:
        - name: id
          in: path
          description: ID to delete
          required: true
          schema:
            type: integer
            format: int64
        - name: pin
          in: path
          description: pin to of the ID to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: pet deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /leaverequest/findByPin:
    get:
      description: Returns all Leave requests identified by the pin
      operationId: find leave requests by pin
      parameters:
        - name: pin
          in: path
          description: pin of user
          required: true
          schema:
            type: integer
            format: int64
        - name: hashIdentityNumber
          in: path
          description: Hashed User Identification number
          required: true
          schema:
            type: string
      responses:
        200:
          description: Leave request response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaveRequest'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    LeaveRequest:
      allOf:
        - $ref: '#/components/schemas/NewLeaveRequest'
        - type: object
          required:
            - id
          properties:
            id:
              type: integer
              format: int64
    NewLeaveRequest:
      type: object
      required:
        # Identify person
        - hashIdentityNumber
        - userPin
        # Identify request Leave
        - startPosition
        - finishPosition
        - reason
        - signature
        - leaveTime
      properties:
        hashIdentityNumber:
          type: string
        userPin:
          type: integer
        startPosition:
          type: string
        finishPosition:
          type: string
        reason:
          type: string
        signature:
          type: string
        leaveTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string